requirements.txt
asgiref==3.11.1
attrs==25.4.0
Django==5.0.3
django-cors-headers==4.3.1
django-filter==24.1
django-unfold==0.90.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.29.0
gunicorn==25.1.0
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
packaging==26.0
pillow==12.1.1
pluggy==1.6.0
psycopg==3.3.3
psycopg-binary==3.3.3
Pygments==2.20.0
PyJWT==2.11.0
PyMySQL==1.1.2
pytest==9.0.3
pytest-django==4.12.0
python-dotenv==1.0.1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
uritemplate==4.2.0
whitenoise==6.12.0